+++ src/main/java/com/mageddo/commons/circuitbreaker/CircuitCheckException.java
@@ -8,4 +8,8 @@ public class CircuitCheckException extends RuntimeException {
   public CircuitCheckException(Throwable e) {
     super(e.getMessage(), e);
   }
+
+  public CircuitCheckException(String message, Throwable e) {
+    super(message, e);
+  }
 }
