+++ src/main/java/com/mageddo/dnsproxyserver/solver/remote/application/ResultSupplier.java
@@ -1,0 +1,9 @@
+package com.mageddo.dnsproxyserver.solver.remote.application;
+
+import com.mageddo.dnsproxyserver.solver.remote.Result;
+
+import java.util.function.Supplier;
+
+public interface ResultSupplier extends Supplier<Result> {
+
+}
