+++ src/main/java/com/mageddo/dnsproxyserver/solver/remote/circuitbreaker/canaryratethreshold/Resilience4jMapper.java
@@ -23,6 +23,7 @@ public class Resilience4jMapper {
 
         .build()
     );
+    circuitBreaker.transitionToOpenState();
     return circuitBreaker;
   }
 }
